Removing api2.FHAuthRequest.  I was able to refactor the original class in a way that doesn't need a new API
